requirements.txt
python-telegram-bot[job-queue]==21.*
pydantic-settings>=2.0
gitpython>=3.1
openai>=1.0
anthropic>=0.28
aiosqlite>=0.20
pexpect>=4.9
